- fix compile warnings and Manager error
svn path=/trunk/boinc/; revision=15905

+++ lib/boinc_cmd.C
@@ -400,7 +400,7 @@ int main(int argc, char** argv) {
                         if (amrr.error_num != ERR_IN_PROGRESS) break;
                         boinc_sleep(1);
                     } else {
-                        unsigned int j, n = amrr.messages.size();
+                        int j, n = (int)amrr.messages.size();
                         if (n) {
                             printf("Messages from account manager:\n");
                             for (j=0; j<n; j++) {
